\name{optrees-package}
\alias{optrees-package}
\alias{optrees}
\docType{package}
\title{Optimal Trees in Weighted Graphs}
\description{
Finds optimal trees in weighted graphs. In particular, this package provides solving tools for minimum cost spanning tree problems, minimum cost arborescence problems, shortest path tree problems and minimum cut tree problem.
}
\details{
\tabular{ll}{
Package: \tab optrees\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-01\cr
License: \tab GPL-3 \cr
}
The most important functions are \link{getMinimumSpanningTree}, \link{getMinimumArborescence}, \link{getShortestPathTree} and \link{getMinimumCutTree}. The other functions included in the package are auxiliary ones that can be used independently.
}
\author{
Manuel Fontenla <manu.fontenla@gmail.com>
}